Fix for the following issues: #132, #121
Fixed the input for exercises. The input for the exercise name must be any string greater than 0 and less than what the box size allows. The exercise can also be a number.
The sets and reps now must be integers greater than 1 of size 1 or 2 digits. The weight must me an integer greater than 1 of size 1, 2, or 3 digits.
There cannot be an empty exercise. Fixed the crash.