Fix for the following issues: #144
Adding programs of the same name was already fixed and referenced in a previous commit. This commit fixes the issue of having multiple exercises with the same name.
You can no longer have two exercises of the same program with the same exact name. You can have any other names you want for the exercises, they can also be numbers if you would like. I don't see it being an issue having numbers in the name of your exercises. I am not expecting a user to want to call their exercise names just standalone numbers because that wouldn't really help you in the gym. But really the user should be able to name the exercise whatever string of chars and integers that they want to help them remember the exercise.