Fix for the following issues: #152, #147, #146, #145, #143, #133, #130, #129, #122, #109, #88, #79, #77, #76, #75, #74, #73, #72
Found a bug when adding a new program when other programs exist. If you were to choose cancel when adding your first exercise for that program, it would remove that program like its supposed to but it wouldnt update the table properly. Now I have it fixed so that it updates properly with the first program in the list and the exercises of the first program populating the table. The exercise database page should now be fully functional. I will reference these issues again if I find other bugs and produce a solution for them.