Fix for the following issues: #114, #124
Removed the feature where the user has to type an exercise name to or program name to remove or edit. Implemented dropdowns inside the alert dialogs. Way easier for users to use.
I tested all the buttons on the Exercise Database page to see if they work properly and if they update the home page properly and found no issues or instances of bugs that cause a crash.