Fix for the following issues: #135, #131, #116, #111, #106
Fixed the login page some. Google sign in button now works and always shows on the login page. Once you do your first google login now, it will remember your account and use that account you used initially only for google sign in. This ensures that the user can only have one account for google sign in.
The progress bar on the login page now works and is only visible and loads when you hit sign in, register, or google sign in and it goes away whenever you gain access to the app or fail login or register.