Fixed the following issues: #154, #153, #84, #83
Current exercise progress bar and current program progress bar work on home page now. The math is correct. You can reset a programs progress by clicking a new program and going back to the program. This is for when you are in the gym and just want to see how much of your workout program you complete in percent form.
Fixed some other stuff on program page also. Added some instructions in the comment block for users. Changed color schemes of alert dialogs and other stuff.